pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tracelens"
version = "1.0.0"
description = "Cloud-native observability analysis system"
requires-python = ">=3.10"
dependencies = [
    "boto3==1.34.30",
    "botocore==1.34.30",
    "python-dateutil==2.9.0.post0",
    "pydantic==2.6.1",
    "fastapi==0.109.2",
    "uvicorn==0.27.1",
    "sentence-transformers==2.6.1",
    "numpy==1.26.4",
    "mangum==0.17.0",
]

[tool.setuptools.packages.find]
where = ["src"]
